Edit: split the view model to Edit and List ViewModels
- this is way better design that the merged one
- using this way, bug of misplaced cursor on configuration changes on
new notes was fixed
- in this way, when a new note is created, it will be inserted to the
database right away,
and then when closing (clearing), it will check that if the note is empty,
it will delete it from the database, if its not empty, the new value
will be recorded to the database
- using the above way means that if there is an old note, and the user
deleted all the contents of it (title and note text) it will be
deleted.
for me I think it's ok to be deleted if it's empty, if the user want it
again, he can just create a new note with that color.